+++ tests/test_admin.py
@@ -1,10 +1,10 @@
 from django.shortcuts import reverse
 from django.test import TestCase
-from wagtail.core.models import Page
-from wagtail.tests.utils import WagtailTestUtils
+from wagtail.models import Page
+from wagtail.test.utils import WagtailTestUtils
 
 from wagtail_trash.models import TrashCan, TrashCanPage
-from wagtail_trash.views import trash_delete
+from wagtail_trash.views import trash_bulk_delete, trash_delete
 from wagtail_trash.wagtail_hooks import TrashCanModelAdmin
 
 trash_admin_url_helper = TrashCanModelAdmin().url_helper
@@ -82,6 +82,60 @@ class TestAdmin(TestCase, WagtailTestUtils):
         assert new_page.child_of(TrashCanPage.objects.first())
         assert TrashCan.objects.count() == 1
 
+    def test_bulk_delete_sets_and_unsets_slug(self):
+        from wagtail_trash.wagtail_hooks import urlconf_time
+
+        root_page = Page.objects.get(url_path="/")
+
+        new_page1 = Page(title="new page 1")
+
+        p1 = root_page.add_child(instance=new_page1)
+
+        p1_slug = p1.slug
+
+        with self.register_hook("before_bulk_action", trash_bulk_delete):
+            delete_url = reverse(
+                "wagtail_bulk_action", args=("wagtailcore", "page", "delete")
+            )
+            id_query = "?id=" + "&id=".join([str(p1.id)])
+            self.client.post(delete_url + id_query)
+
+        p1.refresh_from_db()
+        self.assertEqual(f"trash-{p1.pk}-{p1_slug}", p1.slug)
+
+        with self.register_hook("register_admin_urls", urlconf_time):
+            restore_url = reverse("wagtail_trash_restore", args=(p1.pk,))
+            self.client.get(restore_url)
+
+        p1.refresh_from_db()
+        self.assertEqual(p1_slug, p1.slug)
+
+    def test_bulk_delete_sends_to_trash_can(self):
+        root_page = Page.objects.get(url_path="/")
+
+        new_page1 = Page(title="new page 1")
+        new_page2 = Page(title="new page 2")
+        new_page3 = Page(title="new page 3")
+        new_page4 = Page(title="new page 4")
+
+        p1 = root_page.add_child(instance=new_page1)
+        p2 = root_page.add_child(instance=new_page2)
+        p3 = root_page.add_child(instance=new_page3)
+        p4 = root_page.add_child(instance=new_page4)
+
+        assert TrashCan.objects.count() == 0
+
+        with self.register_hook("before_bulk_action", trash_bulk_delete):
+            delete_url = reverse(
+                "wagtail_bulk_action", args=("wagtailcore", "page", "delete")
+            )
+            id_query = "?id=" + "&id=".join(
+                [str(p1.id), str(p2.id), str(p3.id), str(p4.id)]
+            )
+            self.client.post(delete_url + id_query)
+
+        assert TrashCan.objects.count() == 4
+
     def test_removing_page_from_bin_deletes_it(self):
         root_page = Page.objects.get(url_path="/")
 
@@ -93,11 +147,35 @@ class TestAdmin(TestCase, WagtailTestUtils):
             self.client.post(delete_url)
             assert TrashCan.objects.count() == 1
 
-            delete_url = reverse("wagtailadmin_pages:delete", args=(new_page.id,))
+            trash_can_item = TrashCan.objects.first()
+            delete_url = trash_admin_url_helper.get_action_url(
+                "delete",
+                trash_can_item.id,
+            )
+            self.client.post(delete_url)
+            assert not Page.objects.filter(title="delete page")
+            assert TrashCan.objects.count() == 0
+
+    def test_removing_two_pages_with_same_slug(self):
+        root_page = Page.objects.get(url_path="/")
+
+        cat = Page(title="category 1", slug="category-1")
+        cat_instance = root_page.add_child(instance=cat)
+        cat_2 = Page(title="category 2", slug="category-2")
+        cat_instance_2 = root_page.add_child(instance=cat_2)
+
+        delete = Page(title="delete", slug="delete")
+        delete_instance = cat_instance.add_child(instance=delete)
+        delete_2 = Page(title="delete", slug="delete")
+        delete_instance_2 = cat_instance_2.add_child(instance=delete_2)
+
+        with self.register_hook("before_delete_page", trash_delete):
+            delete_url = reverse("wagtailadmin_pages:delete", args=(delete.id,))
+            self.client.post(delete_url)
+            delete_url = reverse("wagtailadmin_pages:delete", args=(delete_2.id,))
             self.client.post(delete_url)
 
-        assert not Page.objects.filter(title="delete page")
-        assert TrashCan.objects.count() == 0
+            self.assertEqual(TrashCan.objects.count(), 2)
 
     def test_removing_page_unpublishes_all_sub_pages(self):
         root_page = Page.objects.get(url_path="/")