+++ wagtail_trash/managers.py
@@ -1,10 +1,5 @@
-import wagtail
 from django.db import models
-
-if wagtail.VERSION >= (3, 0):
-    from wagtail.query import PageQuerySet
-else:
-    from wagtail.core.query import PageQuerySet
+from wagtail.query import PageQuerySet
 
 from wagtail_trash.models import TrashCanPage