+++ wagtail_trash/models.py
@@ -1,7 +1,7 @@
 from django.contrib.auth import get_user_model
 from django.db import models
 from django.utils.translation import gettext_lazy as _
-from wagtail.core.models import Page
+from wagtail.models import Page
 
 
 class TrashCan(models.Model):