+++ wagtail_trash/utils.py
@@ -1,15 +1,10 @@
 import json
+import re
 
-import wagtail
 from django.core.exceptions import PermissionDenied
 from django.utils.translation import gettext as _
-
-if wagtail.VERSION >= (3, 0):
-    from wagtail.actions.move_page import MovePageAction
-elif wagtail.VERSION >= (2, 16):
-    from wagtail.core.actions.move_page import MovePageAction
-
-from wagtail.core.models import Page, Site
+from wagtail.actions.move_page import MovePageAction
+from wagtail.models import Page, Site
 
 from .models import TrashCanPage
 
@@ -51,13 +46,11 @@ def restore_and_move_page(rb, move_to_page, request):
     if not rb.page.permissions_for_user(request.user).can_move():
         raise PermissionDenied
 
-    if wagtail.VERSION >= (2, 16):
-        action = MovePageAction(
-            rb.page, move_to_page, pos="first-child", user=request.user
-        )
-        action.execute(skip_permission_checks=True)
-    else:
-        rb.page.move(move_to_page, pos="first-child", user=request.user)
+    rb.page.slug = re.sub(r"trash-\d+-", "", rb.page.slug)
+    rb.page.save()
+
+    action = MovePageAction(rb.page, move_to_page, pos="first-child", user=request.user)
+    action.execute(skip_permission_checks=True)
 
     to_be_published_ids = json.loads(rb.data)["published"]