+++ wagtail_trash/views.py
@@ -1,13 +1,10 @@
-import json
-
-import wagtail
 from django.core.exceptions import PermissionDenied
 from django.shortcuts import redirect, render
 from django.utils.http import url_has_allowed_host_and_scheme
 from django.utils.translation import gettext as _
 from treebeard.mp_tree import MP_MoveHandler
 from wagtail.admin import messages
-from wagtail.core.models import Page
+from wagtail.models import Page
 
 from .forms import MoveForm
 from .models import TrashCan
@@ -21,6 +18,49 @@ def get_valid_next_url_from_request(request):
     ):
         return ""
     return next_url
+
+
+def trash_bulk_delete(request, pages):
+    trash_can = trash_can_for_request(request)
+
+    for page in pages:
+        parent = page.get_parent()
+
+        if parent.id == trash_can.id:
+            page.delete(user=request.user)
+
+            messages.success(
+                request, _("Page '{0}' deleted.").format(page.get_admin_display_title())
+            )
+        else:
+            TrashCan.objects.create(
+                page=page,
+                parent=parent,
+                user=request.user,
+                data=generate_page_data(page),
+            )
+
+            page.slug = f"trash-{page.id}-{page.slug}"
+            page.save()
+            page.get_descendants(inclusive=True).unpublish()
+
+            # Preserve the url path
+            old_page = Page.objects.get(id=page.id)
+            new_url_path = old_page.set_url_path(parent=trash_can)
+
+            MP_MoveHandler(page, trash_can, "first-child").process()
+
+            # And reset the url path when in trash
+            new_page = Page.objects.get(id=page.id)
+            new_page.url_path = new_url_path
+            new_page.save()
+
+    next_url = get_valid_next_url_from_request(request)
+
+    if next_url:
+        return redirect(next_url)
+
+    return redirect("wagtailadmin_explore", parent.id)
 
 
 def trash_delete(request, page):
@@ -42,21 +82,20 @@ def trash_delete(request, page):
             page=page, parent=parent, user=request.user, data=generate_page_data(page)
         )
 
+        page.slug = f"trash-{page.id}-{page.slug}"
+        page.save()
         page.get_descendants(inclusive=True).unpublish()
 
-        if wagtail.VERSION >= (2, 16):
-            # Preserve the url path
-            old_page = Page.objects.get(id=page.id)
-            new_url_path = old_page.set_url_path(parent=trash_can)
+        # Preserve the url path
+        old_page = Page.objects.get(id=page.id)
+        new_url_path = old_page.set_url_path(parent=trash_can)
 
-            MP_MoveHandler(page, trash_can, "first-child").process()
+        MP_MoveHandler(page, trash_can, "first-child").process()
 
-            # And reset the url path when in trash
-            new_page = Page.objects.get(id=page.id)
-            new_page.url_path = new_url_path
-            new_page.save()
-        else:
-            page.move(trash_can, pos="first-child", user=request.user)
+        # And reset the url path when in trash
+        new_page = Page.objects.get(id=page.id)
+        new_page.url_path = new_url_path
+        new_page.save()
 
         messages.success(
             request,