+++ wagtail_trash/wagtail_hooks.py
@@ -2,15 +2,15 @@ from django.shortcuts import reverse
 from django.urls import path
 from django.utils.safestring import mark_safe
 from django.utils.translation import gettext as _
+from wagtail import hooks
 from wagtail.contrib.modeladmin.helpers import ButtonHelper, PermissionHelper
 from wagtail.contrib.modeladmin.options import ModelAdmin, modeladmin_register
-from wagtail.contrib.modeladmin.views import IndexView
-from wagtail.core import hooks
-from wagtail.core.models import Page
+from wagtail.contrib.modeladmin.views import DeleteView, IndexView
+from wagtail.models import Page
 
 from .models import TrashCan, TrashCanPage
 from .utils import trash_can_for_request
-from .views import trash_delete, trash_move, trash_restore
+from .views import trash_bulk_delete, trash_delete, trash_move, trash_restore
 
 
 class TrashPermissionHelper(PermissionHelper):
@@ -77,6 +77,21 @@ class TrashCanIndexView(IndexView):
         return _("Trash Can")
 
 
+class TrashCanDeleteView(DeleteView):
+    def confirmation_message(self):
+        return _(
+            "Are you sure you want to delete this %(object)s? If other things in your "
+            "site are related to it, they may also be affected."
+        ) % {"object": _("page")}
+
+    def delete_instance(self):
+        rb = self.instance
+        page = rb.page
+
+        page.delete(user=self.request.user)
+        rb.delete()
+
+
 class TrashCanModelAdmin(ModelAdmin):
     model = TrashCan
     menu_label = _("Trash Can")
@@ -91,6 +106,8 @@ class TrashCanModelAdmin(ModelAdmin):
 
     index_view_class = TrashCanIndexView
     index_template_name = "wagtail_trash/index.html"
+
+    delete_view_class = TrashCanDeleteView
 
     def page_tree(self, rb):
         descendants = rb.page.get_descendants(inclusive=True)
@@ -116,6 +133,14 @@ class TrashCanModelAdmin(ModelAdmin):
 
 
 modeladmin_register(TrashCanModelAdmin)
+
+
+@hooks.register("before_bulk_action")
+def delete_bulk_pages(request, action_type, objects, action_class_instance):
+    from wagtail.admin.views.pages.bulk_actions.delete import DeleteBulkAction
+
+    if action_type == "delete" and isinstance(action_class_instance, DeleteBulkAction):
+        return trash_bulk_delete(request, objects)
 
 
 @hooks.register("before_delete_page")